requirements.txt
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-Cors==3.0.9
Flask-SQLAlchemy==2.4.4
bcrypt==3.2.0
python-dotenv==0.14.0
psycopg2-binary==2.8.6